fixed a bug when compiling with mingw

+++ source/rom.c
@@ -67,7 +67,7 @@ rom* rom_loadEmptyRom(){
     self->type = NES_ROM;
     self->size = 40*1024;
     self->rom_buffer = malloc(self->size);
-    for(int i = 0; i < self->size; i++) self->rom_buffer[i] = 0;
+    for(size_t i = 0; i < self->size; i++) self->rom_buffer[i] = 0;
     return self;
 }
